[손건] 이미지 색상 검색 기능 추가

+++ src/main/java/com/VIG/mvc/service/domain/ImageColor.java
@@ -10,9 +10,9 @@ public class ImageColor implements Serializable{
 	private int colorId;
 	private int ImageId;
 	
-	private float red;
-	private float green;
-	private float blue;
+	private int red;
+	private int green;
+	private int blue;
 	
 	//이미지에서 해당 색이 차지하는 비율
 	private float ratio;
